chromadb_service/requirements.txt
chromadb
sentence-transformers
grpcio
grpcio-tools
